+++ fluxc/src/main/java/org/wordpress/android/fluxc/module/ReleaseNetworkModule.java
@@ -36,6 +36,7 @@ import org.wordpress.android.fluxc.network.rest.wpcom.post.PostRestClient;
 import org.wordpress.android.fluxc.network.rest.wpcom.reader.ReaderRestClient;
 import org.wordpress.android.fluxc.network.rest.wpcom.site.SiteRestClient;
 import org.wordpress.android.fluxc.network.rest.wpcom.stats.InsightsRestClient;
+import org.wordpress.android.fluxc.network.rest.wpcom.stats.time.StatsUtils;
 import org.wordpress.android.fluxc.network.rest.wpcom.stockmedia.StockMediaRestClient;
 import org.wordpress.android.fluxc.network.rest.wpcom.taxonomy.TaxonomyRestClient;
 import org.wordpress.android.fluxc.network.rest.wpcom.theme.ThemeRestClient;
@@ -191,11 +192,12 @@ public class ReleaseNetworkModule {
     @Singleton
     @Provides
     public InsightsRestClient provideInsightsRestClient(Context appContext, Dispatcher dispatcher,
-                                                     @Named("regular") RequestQueue requestQueue,
-                                                     AccessToken token, UserAgent userAgent,
-                                                     WPComGsonRequestBuilder wpComGsonRequestBuilder) {
+                                                        @Named("regular") RequestQueue requestQueue,
+                                                        AccessToken token, UserAgent userAgent,
+                                                        WPComGsonRequestBuilder wpComGsonRequestBuilder,
+                                                        StatsUtils statsUtils) {
         return new InsightsRestClient(dispatcher, wpComGsonRequestBuilder, appContext, requestQueue, token,
-                userAgent);
+                userAgent, statsUtils);
     }
 
     @Singleton
