+++ fluxc/src/main/java/org/wordpress/android/fluxc/network/utils/StatsGranularity.kt
@@ -1,0 +1,12 @@
+package org.wordpress.android.fluxc.network.utils
+
+enum class StatsGranularity(private val value: String) {
+    DAYS("day"),
+    WEEKS("week"),
+    MONTHS("month"),
+    YEARS("year");
+
+    override fun toString(): String {
+        return value
+    }
+}
